refactor: generic args/result
